+++ Python/server/__init__.py
@@ -66,6 +66,7 @@ def bootstrap():
     from server import localization_tools           # noqa: F401  Sub-batch V
     from server import testing_validation_tools     # noqa: F401  Sub-batch W
     from server import data_table_extension_tools   # noqa: F401  Sub-batch X
+    from server import metasound_tools              # noqa: F401  Sub-batch Y
     from server import packaging_tools             # noqa: F401
     from server import vertical_test_tools        # noqa: F401